+++ CHANGELOG.md
@@ -56,11 +56,15 @@ voir si on peut pas modifier les coupes par des intersection par une boite.
 voir si plusieurs MSEC dans un lattice est reconnu par VEGETA
 
 
-v 89 11 07 23 
-- mesh z placement corrected
+v 90 12 07 23
 - moret : lattice reading and creation simplified
 - moret : disb reading added.
 - moret : create_one_lattice simplification
+- moret : lattice"s parent substraction moved to moretMeshTransformer.
+
+v 89 11 07 23 
+- mesh z placement corrected
+
 
 v 88 14 06 23
 - z-cut corrected.